requirements.txt

django==2.2.1
django-simpleui==2.1
djangorestframework==3.9.1
xlwt
lxml
pyecharts

imutils
pymysql
